250116_업데이트

+++ MapleContents/CerniumPlazaMode.h
@@ -75,7 +75,7 @@ private:
 	float SpeedMultiplier = 0.0f;
 	float PreviousPlayerLocation = 0.0f;
 
-	double Velocity = 0.0f;
+	float Velocity = 0.0f;
 
 };
