+++ MapleContents/MapleStruct.h
@@ -3,6 +3,6 @@
 
 struct PlayerStatus
 {
-	int Hp = 100.0f;
-	int Att = 1.0f;
+	float Hp = 100.0f;
+	float Att = 1.0f;
 };
